pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-service-monitor"
version = "0.1.0"
description = "AI-Powered Service Health Monitor with LLM analysis and predictive scaling"
authors = [
    {name = "AI Monitor Team", email = "team@ai-monitor.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "langchain>=0.1.0",
    "langchain-community>=0.2.0",
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "prometheus-client>=0.19.0",
    "kafka-python>=2.0.0",
    "redis>=5.0.0",
    "kubernetes>=28.1.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "numpy>=1.25.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0.0",
    "python-json-logger>=2.0.0",
    "structlog>=23.2.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
viz = [
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]